alıntı ekleme

+++ src/Components/WordsPage/WordsPage.js
@@ -5,6 +5,7 @@ import { FaRegTrashAlt } from 'react-icons/fa';
 import VerticalNavbar from '../VerticalNavbar/VerticalNavbar';
 import axios from 'axios';
 import { useCookies } from 'react-cookie';
+import Swal1 from 'sweetalert2'
 
 function WordsPage() {
   const [cookies] = useCookies(['email']);
@@ -41,19 +42,39 @@ function WordsPage() {
   };
   const title =null;
   const value =null;
+
+  
   const handleDelete = (item, word) => {
- 
-    //console.log(word); `http://127.0.0.1:8000/api/words/${email}/${word}`
-    axios.delete(`https://librarygop.com/public/index.php/api/words_delete/${email}/${word}`)
-    .then(response => {
-      // İstek başarılı olduğunda yapılacak işlemler
-      console.log("Kelime silindi:", response.data);
-      window.location.reload();
+    Swal1.fire({
+      title: 'Emin Misiniz?',
+      text: "Bu kelimeyi silmek istediğinizden emin misiniz?!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#1c8b78',
+      cancelButtonColor: 'gray',
+      confirmButtonText: 'Evet!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        axios.delete(`https://librarygop.com/public/index.php/api/words_delete/${email}/${word}`)
+        .then(response => {
+          // İstek başarılı olduğunda yapılacak işlemler
+          console.log("Kelime silindi:", response.data);
+          window.location.reload();
+        
+        })
+        .catch(error => {
+          // Hata durumunda yapılacak işlemler
+          console.error("bir hata oluştu:", error);
+        });
+    
+        Swal1.fire(
+          'silindi!',
+          'Kelime Silindi.',
+          'success'
+        )
+      }
     })
-    .catch(error => {
-      // Hata durumunda yapılacak işlemler
-      console.error("bir hata oluştu:", error);
-    });
+    //console.log(word); `http://127.0.0.1:8000/api/words/${email}/${word}`
     
   };
